Rename test/core to test/renderer
