+++ test/unit/renderer/schema.test.js
