+++ test/unit/renderer/viz/colorspaces.test.js
