+++ test/unit/renderer/viz/expressions/Fade.test.js
