+++ test/unit/renderer/viz/expressions/aggregation/clusterAggregation.test.js
