+++ test/unit/renderer/viz/expressions/aggregation/globalAggregation.test.js
