+++ test/unit/renderer/viz/expressions/blend.test.js
