+++ test/unit/renderer/viz/expressions/color/hex.test.js
