+++ test/unit/renderer/viz/expressions/color/named-color.test.js
