+++ test/unit/renderer/viz/expressions/color/opacity.test.js
