+++ test/unit/renderer/viz/expressions/color/palettes.test.js
