+++ test/unit/renderer/viz/expressions/color/rgb.test.js
