+++ test/unit/renderer/viz/expressions/sprite.test.js
