+++ test/unit/renderer/viz/expressions/sprites.test.js
